Fixing multi-line description in PreviewCard. (#18)
Either supply a description text, or not at all in the whole PreviewCard cluster.
Even if we supply a description text, it will be limited to one line, because it doesn't look good and I don't want to resort to hacky solution when compromising between single or multi-line description.
Then a further limitation, we can't use overflow: hidden + whitespace: nowrap + text-overflow: ellipsis combo in here, because flex determine that the full length of the text given will be used. So non-Chrome user will be limited to slightly worse experience, but it doesn't really matter.